+++ questdb-rs/src/tests/mock.rs
@@ -37,7 +37,7 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use crate::ingress;
-#[cfg(feature = "ilp-over-http")]
+#[cfg(feature = "sync-sender-http")]
 use std::io::Write;
 
 use super::ndarr::ArrayColumnTypeTag;
@@ -54,7 +54,7 @@ pub struct MockServer {
     pub host: &'static str,
     pub port: u16,
     pub msgs: Vec<Vec<u8>>,
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     settings_response: serde_json::Value,
 }
 
@@ -84,7 +84,7 @@ fn tls_config() -> Arc<ServerConfig> {
     Arc::new(config)
 }
 
-#[cfg(feature = "ilp-over-http")]
+#[cfg(feature = "sync-sender-http")]
 pub struct HttpRequest {
     method: String,
     path: String,
@@ -92,7 +92,7 @@ pub struct HttpRequest {
     body: Vec<u8>,
 }
 
-#[cfg(feature = "ilp-over-http")]
+#[cfg(feature = "sync-sender-http")]
 impl HttpRequest {
     pub fn method(&self) -> &str {
         &self.method
@@ -111,7 +111,7 @@ impl HttpRequest {
     }
 }
 
-#[cfg(feature = "ilp-over-http")]
+#[cfg(feature = "sync-sender-http")]
 pub struct HttpResponse {
     status_code: u16,
     status_text: String,
@@ -119,7 +119,7 @@ pub struct HttpResponse {
     body: Vec<u8>,
 }
 
-#[cfg(feature = "ilp-over-http")]
+#[cfg(feature = "sync-sender-http")]
 impl HttpResponse {
     pub fn empty() -> Self {
         HttpResponse {
@@ -181,14 +181,14 @@ impl HttpResponse {
     }
 }
 
-#[cfg(feature = "ilp-over-http")]
+#[cfg(feature = "sync-sender-http")]
 fn contains(haystack: &[u8], needle: &[u8]) -> bool {
     haystack
         .windows(needle.len())
         .any(|window| window == needle)
 }
 
-#[cfg(feature = "ilp-over-http")]
+#[cfg(feature = "sync-sender-http")]
 fn position(haystack: &[u8], needle: &[u8]) -> Option<usize> {
     haystack
         .windows(needle.len())
@@ -211,7 +211,7 @@ impl MockServer {
             host: "localhost",
             port,
             msgs: Vec::new(),
-            #[cfg(feature = "ilp-over-http")]
+            #[cfg(feature = "sync-sender-http")]
             settings_response: serde_json::Value::Null,
         })
     }
@@ -294,7 +294,7 @@ impl MockServer {
         self.wait_for(timeout, |event| event.is_readable())
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     pub fn wait_for_send(&mut self, duration: Option<Duration>) -> io::Result<bool> {
         self.wait_for(duration, |event| event.is_writable())
     }
@@ -309,7 +309,7 @@ impl MockServer {
         }
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     pub fn configure_settings_response(
         mut self,
         supported_versions: &[u16],
@@ -332,7 +332,7 @@ impl MockServer {
         self
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     fn do_write(&mut self, buf: &[u8]) -> io::Result<usize> {
         let client = self.client.as_mut().unwrap();
         if let Some(tls_conn) = self.tls_conn.as_mut() {
@@ -343,7 +343,7 @@ impl MockServer {
         }
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     fn do_write_all(&mut self, buf: &[u8], timeout_sec: Option<f64>) -> io::Result<()> {
         let deadline = timeout_sec.map(|sec| Instant::now() + Duration::from_secs_f64(sec));
         let mut pos = 0usize;
@@ -378,7 +378,7 @@ impl MockServer {
         Ok(())
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     fn read_more(&mut self, accum: &mut Vec<u8>, deadline: Instant, stage: &str) -> io::Result<()> {
         let mut chunk = [0u8; 1024];
         let count = match self.do_read(&mut chunk[..]) {
@@ -415,7 +415,7 @@ impl MockServer {
         Ok(())
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     fn recv_http_method(
         &mut self,
         accum: &mut Vec<u8>,
@@ -444,7 +444,7 @@ impl MockServer {
         Ok((body_start, method, path))
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     fn recv_http_headers(
         &mut self,
         pos: usize,
@@ -473,7 +473,7 @@ impl MockServer {
         Ok((body_start, headers))
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     pub fn send_http_response(
         &mut self,
         response: HttpResponse,
@@ -483,7 +483,7 @@ impl MockServer {
         Ok(())
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     pub fn send_settings_response(&mut self) -> io::Result<()> {
         let response = HttpResponse::empty()
             .with_status(200, "OK")
@@ -492,12 +492,12 @@ impl MockServer {
         Ok(())
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     pub fn send_http_response_q(&mut self, response: HttpResponse) -> io::Result<()> {
         self.send_http_response(response, Some(5.0))
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     pub fn recv_http(&mut self, wait_timeout_sec: f64) -> io::Result<HttpRequest> {
         let mut accum = Vec::<u8>::new();
         let deadline = Instant::now() + Duration::from_secs_f64(wait_timeout_sec);
@@ -530,7 +530,7 @@ impl MockServer {
         })
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     pub fn recv_http_q(&mut self) -> io::Result<HttpRequest> {
         self.recv_http(5.0)
     }
@@ -625,12 +625,12 @@ impl MockServer {
         SenderBuilder::new(Protocol::Tcps, self.host, self.port)
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     pub fn lsb_http(&self) -> SenderBuilder {
         SenderBuilder::new(Protocol::Http, self.host, self.port)
     }
 
-    #[cfg(feature = "ilp-over-http")]
+    #[cfg(feature = "sync-sender-http")]
     pub fn lsb_https(&self) -> SenderBuilder {
         SenderBuilder::new(Protocol::Https, self.host, self.port)
     }
